LayoutTests/http/tests/xmlhttprequest/cross-origin-authorization-with-embedder.html

<button onclick="presetAuthorization()">Start</button>
<pre id="console"></pre>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
    layoutTestController.setCanOpenWindows();
}

function log(message)
{
    document.getElementById('console').appendChild(document.createTextNode(message + '\n'));
}

function presetAuthorization()
{
    if (window.layoutTestController)
        layoutTestController.authenticateSession("http://localhost:8000/xmlhttprequest/resources/cross-origin-authorization.php", "test", "test");

    test();
}

function test()
{
    var req = new XMLHttpRequest;
    req.open("GET", "http://localhost:8000/xmlhttprequest/resources/cross-origin-authorization.php", false);
    try {
        req.send();
        log((req.status == 401) ? "PASS" : "FAIL: credentials were sent");
    } catch (ex) {
        log("FAIL: Got an exception. " + ex);
    }
    log("DONE");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}


if (window.layoutTestController)
    presetAuthorization();
</script>
