LayoutTests/platform/mac-tiger/http/tests/xmlhttprequest/cross-origin-authorization-with-embedder-expected.txt

CONSOLE MESSAGE: line 1: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/cross-origin-authorization.php. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Start
PASS
DONE
